JT - update data structure to allow fit time to be a subset of all data
In some cases (e.g. retrospective studies), the user may want to plot all available observed data from a season, but only fit and forecast a subset of the data.  This update separates 'all' data into the $data list entry and the data to be fit into the $data_fit entry.